Create todo_v2.cs
This is an updated version which will allow for new users and better error handling. this is the first version the user is able to create an account and generate tasks however ...

The user can login with any username.
tasks are shared with every user no private lists.
needs a whole lot more work.